BIG UPDATE

+++ src/main/java/spring_boot/project_swp/dto/request/BookingRequest.java
@@ -2,15 +2,10 @@ package spring_boot.project_swp.dto.request;
 
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import java.time.LocalDateTime;
+import lombok.*;
 import lombok.experimental.FieldDefaults;
 import spring_boot.project_swp.entity.BookingTypeEnum;
-
-import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
@@ -18,21 +13,20 @@ import java.time.LocalDateTime;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BookingRequest {
-    @NotNull(message = "User ID cannot be null")
-    Integer userId;
+  @NotNull(message = "User ID cannot be null")
+  Long userId;
 
-    @NotNull(message = "Vehicle ID cannot be null")
-    Integer vehicleId;
+  @NotNull(message = "Vehicle ID cannot be null")
+  Long vehicleId;
 
-    @NotNull(message = "Booking type cannot be null")
-    BookingTypeEnum bookingType;
+  @NotNull(message = "Booking type cannot be null")
+  BookingTypeEnum bookingType;
 
-    @NotNull(message = "Start time cannot be null")
-    @FutureOrPresent(message = "Start time must be in the present or future")
-    LocalDateTime startTime;
+  @NotNull(message = "Start time cannot be null")
+  @FutureOrPresent(message = "Start time must be in the present or future")
+  LocalDateTime startTime;
 
-    @NotNull(message = "End time cannot be null")
-    @FutureOrPresent(message = "End time must be in the present or future")
-    LocalDateTime endTime;
-
+  @NotNull(message = "End time cannot be null")
+  @FutureOrPresent(message = "End time must be in the present or future")
+  LocalDateTime endTime;
 }
