+++ src/main/java/spring_boot/project_swp/dto/response/BookingResponse.java
@@ -1,5 +1,6 @@
 package spring_boot.project_swp.dto.response;
 
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,9 +9,6 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import spring_boot.project_swp.entity.BookingStatusEnum;
 import spring_boot.project_swp.entity.BookingTypeEnum;
-import spring_boot.project_swp.dto.response.VehicleResponse;
-
-import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
@@ -18,14 +16,14 @@ import java.time.LocalDateTime;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BookingResponse {
-    Integer bookingId;
-    Integer userId;
-    String userName;
-    VehicleResponse vehicle;
-    BookingTypeEnum bookingType;
-    LocalDateTime startTime;
-    LocalDateTime endTime;
-    Double totalAmount;
-    BookingStatusEnum status;
-    LocalDateTime createdAt;
+  Long bookingId;
+  Long userId;
+  String userName;
+  VehicleResponse vehicle;
+  BookingTypeEnum bookingType;
+  LocalDateTime startTime;
+  LocalDateTime endTime;
+  Double totalAmount;
+  BookingStatusEnum status;
+  LocalDateTime createdAt;
 }
