+++ src/main/java/spring_boot/project_swp/dto/response/DiscountResponse.java
@@ -1,14 +1,13 @@
 package spring_boot.project_swp.dto.response;
 
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.experimental.FieldDefaults;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
 
 @Data
 @NoArgsConstructor
@@ -17,16 +16,16 @@ import java.time.LocalDateTime;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DiscountResponse {
 
-    Integer discountId;
-    String code;
-    String description;
-    BigDecimal amountPercentage;
-    BigDecimal amountFixed;
-    LocalDateTime startDate;
-    LocalDateTime endDate;
-    Integer minRentalDuration;
-    BigDecimal maxDiscountAmount;
-    Integer usageLimit;
-    Integer currentUsage;
-    Boolean isActive;
+  Long discountId;
+  String code;
+  String description;
+  BigDecimal amountPercentage;
+  BigDecimal amountFixed;
+  LocalDateTime startDate;
+  LocalDateTime endDate;
+  Integer minRentalDuration;
+  BigDecimal maxDiscountAmount;
+  Integer usageLimit;
+  Integer currentUsage;
+  Boolean isActive;
 }
