+++ src/main/java/spring_boot/project_swp/dto/response/LocationResponse.java
@@ -1,24 +1,23 @@
 package spring_boot.project_swp.dto.response;
 
+import java.math.BigDecimal;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class LocationResponse {
-    int locationId;
-    String locationName;
-    String locationType;
-    BigDecimal latitude;
-    BigDecimal longitude;
-    BigDecimal radius;
-    LocationResponse parent;
-    boolean active;
+  Long locationId;
+  String locationName;
+  String locationType;
+  BigDecimal latitude;
+  BigDecimal longitude;
+  BigDecimal radius;
+  LocationResponse parent;
+  boolean active;
 }
