+++ src/main/java/spring_boot/project_swp/dto/response/RentalResponse.java
@@ -1,5 +1,6 @@
 package spring_boot.project_swp.dto.response;
 
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,32 +8,30 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
-import java.time.LocalDateTime;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class RentalResponse {
-    Integer rentalId;
-    Integer bookingId;
-    Integer renterId;
-    String renterName;
-    Integer vehicleId;
-    String vehicleModel;
-    Integer pickupStationId;
-    String pickupStationName;
-    Integer returnStationId;
-    String returnStationName;
-    Integer pickupStaffId;
-    String pickupStaffName;
-    Integer returnStaffId;
-    String returnStaffName;
-    LocalDateTime startTime;
-    LocalDateTime endTime;
-    Double totalCost;
-    String status;
-    String contractUrl;
-    LocalDateTime createdAt;
+  Long rentalId;
+  Long bookingId;
+  Long renterId;
+  String renterName;
+  Long vehicleId;
+  String vehicleModel;
+  Long pickupStationId;
+  String pickupStationName;
+  Long returnStationId;
+  String returnStationName;
+  Long pickupStaffId;
+  String pickupStaffName;
+  Long returnStaffId;
+  String returnStaffName;
+  LocalDateTime startTime;
+  LocalDateTime endTime;
+  Double totalCost;
+  String status;
+  String contractUrl;
+  LocalDateTime createdAt;
 }
