+++ src/main/java/spring_boot/project_swp/dto/response/VehicleModelResponse.java
@@ -11,10 +11,10 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class VehicleModelResponse {
-    private int modelId;
-    private String modelName;
-    private String brand;
-    private String type;
-    private int capacityKWh;
-    private String description;
+  private Long modelId;
+  private String modelName;
+  private String brand;
+  private String type;
+  private int capacityKWh;
+  private String description;
 }
