+++ src/main/java/spring_boot/project_swp/dto/response/VehicleResponse.java
@@ -11,26 +11,27 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class VehicleResponse {
-    private int vehicleId;
-    private String licensePlate;
-    private int batteryCapacity;
-    private int currentBattery;
-    private String vehicleStatus;
-    private double pricePerHour;
-    private ModelInfo model;
-    private StationInfo currentStation;
-    private String imageUrl;
+  private Long vehicleId;
+  private String licensePlate;
+  private int batteryCapacity;
+  private int currentBattery;
+  private String vehicleStatus;
+  private double pricePerHour;
+  private ModelInfo model;
+  private StationInfo currentStation;
+  private String imageUrl;
 
-    @Data
-    public static class ModelInfo {
-        private int modelId;
-        private String modelName;
-        private String brand;
-    }
-    @Data
-    public static class StationInfo {
-        private int stationId;
-        private String stationName;
-        private String address;
-    }
+  @Data
+  public static class ModelInfo {
+    private Long modelId;
+    private String modelName;
+    private String brand;
+  }
+
+  @Data
+  public static class StationInfo {
+    private Long stationId;
+    private String stationName;
+    private String address;
+  }
 }
