+++ src/main/java/spring_boot/project_swp/entity/Location.java
@@ -2,6 +2,10 @@ package spring_boot.project_swp.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,11 +13,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.annotations.CreationTimestamp;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "Locations")
@@ -23,42 +22,46 @@ import java.util.List;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Location {
-    @Id
-    @Column(name = "LocationId")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer locationId;
+  @Id
+  @Column(name = "LocationId")
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  Long locationId;
 
-    @Column(name = "LocationName", nullable = false, length = 100, columnDefinition = "NVARCHAR(100)")
-    String locationName;
+  @Column(name = "LocationName", nullable = false, length = 100, columnDefinition = "NVARCHAR(100)")
+  String locationName;
 
-    @Column(name = "LocationType", nullable = false, length = 50, columnDefinition = "NVARCHAR(50)")
-    String locationType;
+  @Column(name = "LocationType", nullable = false, length = 50, columnDefinition = "NVARCHAR(50)")
+  String locationType;
 
-    @Column(name = "Latitude", precision = 10, scale = 2)
-    BigDecimal latitude;
+  @Column(name = "Latitude", precision = 10, scale = 2)
+  BigDecimal latitude;
 
-    @Column(name = "Longitude", precision = 10, scale = 2)
-    BigDecimal longitude;
+  @Column(name = "Longitude", precision = 10, scale = 2)
+  BigDecimal longitude;
 
-    @Column(name = "Radius", precision = 5, scale = 2)
-    BigDecimal radius;
+  @Column(name = "Radius", precision = 5, scale = 2)
+  BigDecimal radius;
 
-    @Column(name = "IsActive", nullable = false)
-    boolean isActive;
+  @Column(name = "IsActive", nullable = false)
+  boolean isActive;
 
-    @CreationTimestamp
-    @Column(name = "CreatedAt", nullable = false, updatable = false)
-    LocalDateTime createdAt;
+  @CreationTimestamp
+  @Column(name = "CreatedAt", nullable = false, updatable = false)
+  LocalDateTime createdAt;
 
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "ParentLocationId")
-    Location parent;
+  @ManyToOne(fetch = FetchType.EAGER)
+  @JoinColumn(name = "ParentLocationId")
+  Location parent;
 
-    @OneToMany(mappedBy = "parent", fetch = FetchType.LAZY, targetEntity = Location.class)
-    @JsonIgnore // Thêm annotation để tránh vòng lặp vô hạn khi serialize
-    List<Location> children = new ArrayList<>();
+  @OneToMany(mappedBy = "parent", fetch = FetchType.LAZY, targetEntity = Location.class)
+  @JsonIgnore
+  List<Location> children = new ArrayList<>();
 
-    @OneToMany(mappedBy = "location", fetch = FetchType.EAGER, cascade = CascadeType.ALL, orphanRemoval = true)
-    @JsonIgnore
-    List<Station> stations = new ArrayList<>();
+  @OneToMany(
+      mappedBy = "location",
+      fetch = FetchType.EAGER,
+      cascade = CascadeType.ALL,
+      orphanRemoval = true)
+  @JsonIgnore
+  List<Station> stations = new ArrayList<>();
 }
