+++ src/main/java/spring_boot/project_swp/entity/RentalDiscounts.java
@@ -1,16 +1,14 @@
 package spring_boot.project_swp.entity;
 
 import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.experimental.FieldDefaults;
-
-import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
 
 @Entity
 @Table(name = "RentalDiscounts")
@@ -22,19 +20,19 @@ import java.time.LocalDateTime;
 @IdClass(RentalDiscountsId.class)
 public class RentalDiscounts {
 
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "RentalId", referencedColumnName = "RentalId")
-    Rental rental;
+  @Id
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "RentalId", referencedColumnName = "RentalId")
+  Rental rental;
 
-    @Id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "DiscountId", referencedColumnName = "DiscountId")
-    Discount discount;
+  @Id
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "DiscountId", referencedColumnName = "DiscountId")
+  Discount discount;
 
-    @Column(name = "AppliedAmount", precision = 10, scale = 2)
-    int appliedAmount;
+  @Column(name = "AppliedAmount", precision = 10, scale = 2)
+  BigDecimal appliedAmount;
 
-    @Column(name = "AppliedAt")
-    LocalDateTime appliedAt;
+  @Column(name = "AppliedAt")
+  LocalDateTime appliedAt;
 }
