+++ src/main/java/spring_boot/project_swp/entity/RentalDiscountsId.java
@@ -1,15 +1,14 @@
 package spring_boot.project_swp.entity;
 
+import java.io.Serializable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class RentalDiscountsId implements Serializable {
-    private Integer rental;
-    private Integer discount;
+  private Long rental;
+  private Long discount;
 }
